requirements.txt
# ── Telegram ────────────────────────────────────────────────
python-telegram-bot[webhooks,job-queue]==21.7
pyrogram==2.0.106
tgcrypto==1.2.5
telethon==1.41.2
apscheduler~=3.10.4

# ── Database: PostgreSQL ────────────────────────────────────
sqlalchemy==1.4.46
psycopg2-binary==2.9.10

# ── Database: MongoDB ───────────────────────────────────────
pymongo==4.10.1
motor==3.7.1

# ── HTTP / async ────────────────────────────────────────────
aiohttp==3.11.10
aiofiles>=24.1.0
requests==2.32.3

# ── Image processing ────────────────────────────────────────
Pillow==11.1.0

# ── PDF ─────────────────────────────────────────────────────
fpdf2==2.8.3

# ── HTML / parsing ──────────────────────────────────────────
beautifulsoup4==4.12.3
bs4==0.0.2
lxml>=4.6.5,<5
bleach==6.3.0

# ── Translation ─────────────────────────────────────────────
gpytranslate>=2.0.0
deep-translator==1.11.4

# ── Search / Wikipedia ──────────────────────────────────────
wikipedia==1.4.0

# ── Rate limiting ───────────────────────────────────────────
pyrate-limiter==2.10.0

# ── Markdown ────────────────────────────────────────────────
markdown2==2.5.3

# ── Telegraph ───────────────────────────────────────────────
telegraph==2.1.0

# ── Speed test ──────────────────────────────────────────────
speedtest-cli==2.1.3

# ── File analysis ───────────────────────────────────────────
hachoir==3.3.0

# ── System ──────────────────────────────────────────────────
psutil==6.1.0
python-dotenv==1.0.1
cachetools==4.2.2
countryinfo==0.1.2
emoji==2.15.0
regex==2025.10.23
